backend/requirements.txt
django
djangorestframework
django-filter
djangorestframework-simplejwt
django-channels
django-cors-headers
pillow